linting, vault functions

+++ contracts/vault/Vault.sol
@@ -8,50 +8,54 @@ import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
 contract Vault is Pausable, Ownable {
     using SafeERC20 for IERC20;
-    
-    uint256 timelock = 137092276;   // 4 years, 4 months, 4 days ...
+
+    uint256 timelock = 137092276; // 4 years, 4 months, 4 days ...
     uint256 rewardsPerDay;
 
     // map a users stakes to a specific pool
     mapping(address => mapping(address => Stake[])) Stakes;
 
     struct Stake {
-        uint256 amount;          // quantity staked
-        uint256 startBlock;      // stake creation timestamp
-        uint256 timeLockEnd;     // The point at which the (4 yr, 4 mo, 4 day) timelock ends for a stake, and thus the funds can be withdrawn.
-        bool active;             // true = stake in vault, false = user withdrawn stake
+        uint256 amount; // quantity staked
+        uint256 startBlock; // stake creation timestamp
+        uint256 timeLockEnd; // The point at which the (4 yr, 4 mo, 4 day) timelock ends for a stake, and thus the funds can be withdrawn.
+        bool active; // true = stake in vault, false = user withdrawn stake
     }
 
     struct UserPosition {
-        uint256 totalAmount;     // total value staked by user in given pool
-        uint256 rewardDebt;      // house fee (?)
-        uint256[] sKey;          // list of user stakes in pool subject to timelock
-        bool staticLock;         // guarantees a users stake is locked, even after timelock expiration
-        bool autocompounding;    // this userposition enables auto compounding (Auto restaking the rewards)
+        uint256 totalAmount; // total value staked by user in given pool
+        uint256 rewardDebt; // house fee (?)
+        uint256[] sKey; // list of user stakes in pool subject to timelock
+        bool staticLock; // guarantees a users stake is locked, even after timelock expiration
+        bool autocompounding; // this userposition enables auto compounding (Auto restaking the rewards)
     }
 
     // user position tracking
     mapping(address => mapping(address => UserPosition)) UserPositions; // account => (token => userposition)
 
     struct Pool {
-        uint256 totalPooled;        // total token pooled in the contract
-        uint256 rewardsPerBlock;    // rate at which CAPL is minted for this pool
-        uint256 accCaplPerShare;    // weighted CAPL share in pool
-        uint256 lastRewardBlock;    // last time a claim was made
+        uint256 totalPooled; // total token pooled in the contract
+        uint256 rewardsPerBlock; // rate at which CAPL is minted for this pool
+        uint256 accCaplPerShare; // weighted CAPL share in pool
+        uint256 lastRewardBlock; // last time a claim was made
     }
 
     // pool tracking
     mapping(address => Pool) Pools; // token => pool
     mapping(address => UserPosition[]) PoolUsers; // token => userPosition
 
-    event DepositVault(address user, address token, uint amount);
-    event WithdrawVault(address user, address token, uint amount);
-    event WithdrawMATIC(address destination, uint amount);
+    event DepositVault(address user, address token, uint256 amount);
+    event WithdrawVault(address user, address token, uint256 amount);
+    event WithdrawMATIC(address destination, uint256 amount);
 
     // TBD: Assume creation with one pool required (?)
-    constructor () {}
-    
-    function depositVault(address _token, address _user, uint256 _amount) external whenNotPaused {
+    constructor() {}
+
+    function depositVault(
+        address _token,
+        address _user,
+        uint256 _amount
+    ) external whenNotPaused {
         require(_amount > 0, "Amount 0");
 
         // userPosition
@@ -67,7 +71,7 @@ contract Vault is Pausable, Ownable {
             require(!checkIfPoolExists(_token), "This pool already exists.");
 
             // create new stake
-            Stake memory newStake = Stake ({
+            Stake memory newStake = Stake({
                 amount: _amount,
                 startBlock: block.timestamp,
                 timeLockEnd: block.timestamp + timelock,
@@ -97,12 +101,20 @@ contract Vault is Pausable, Ownable {
         emit DepositVault(_user, _token, _amount);
     }
 
-    function withdrawVault(address _user, address _token, uint256 _amount, uint _stakeId) external whenNotPaused {
+    function withdrawVault(
+        address _user,
+        address _token,
+        uint256 _amount,
+        uint256 _stakeId
+    ) external whenNotPaused {
         require(_amount > 0, "Amount 0");
 
         // we should consider about the withdrawfee during the timelock
         UserPosition storage userPosition = UserPositions[_user][_token];
-        require(userPosition.totalAmount > _amount, "Withdrawn amount exceed the user balance");
+        require(
+            userPosition.totalAmount > _amount,
+            "Withdrawn amount exceed the user balance"
+        );
         userPosition.totalAmount -= _amount;
 
         // Update Pool info
@@ -121,7 +133,7 @@ contract Vault is Pausable, Ownable {
         IERC20(_token).safeTransferFrom(address(this), _user, _amount);
         emit WithdrawVault(_user, _token, _amount);
     }
-   
+
     /*
         Read functions
     */
@@ -135,13 +147,16 @@ contract Vault is Pausable, Ownable {
     function checkIfPoolExists(address _token) public view returns (bool) {
         return Pools[_token].totalPooled > 0;
     }
-    
+
     /*
         Admin functions
         TODO: Add RBAC for all
     */
 
-    function addPool(address _token, uint256 _rewardsPerBlock) external onlyOwner {
+    function addPool(address _token, uint256 _rewardsPerBlock)
+        external
+        onlyOwner
+    {
         require(!checkIfPoolExists(_token), "This pool already exists.");
 
         Pool memory pool = Pool({
@@ -153,11 +168,16 @@ contract Vault is Pausable, Ownable {
 
         Pools[_token] = pool;
     }
-    
-    function addUserPosition(address _token, address _user, uint256 _amount, uint _rewardDebt) public onlyOwner {
+
+    function addUserPosition(
+        address _token,
+        address _user,
+        uint256 _amount,
+        uint256 _rewardDebt
+    ) public onlyOwner {
         uint256[] memory userStakeKeys;
 
-        UserPosition memory newUser = UserPosition ({
+        UserPosition memory newUser = UserPosition({
             totalAmount: _amount,
             rewardDebt: _rewardDebt,
             sKey: userStakeKeys,
@@ -172,14 +192,22 @@ contract Vault is Pausable, Ownable {
     /*  This function will check if a new stake needs to be created based on lockingThreshold.
         See readme for details.
     */
-    function checkTimelockThreshold(Stake storage _lastStake) internal view returns (bool) {
+    function checkTimelockThreshold(Stake storage _lastStake)
+        internal
+        view
+        returns (bool)
+    {
         return _lastStake.timeLockEnd < block.timestamp;
     }
 
-    function addStake(address _token, address _user, uint256 _amount) external {
+    function addStake(
+        address _token,
+        address _user,
+        uint256 _amount
+    ) external {
         // create user & stake data
         Stake memory stake = Stake({
-            amount: _amount,                   // first stake
+            amount: _amount, // first stake
             startBlock: block.timestamp,
             timeLockEnd: block.timestamp + timelock,
             active: true
@@ -188,20 +216,26 @@ contract Vault is Pausable, Ownable {
         Stakes[_user][_token].push(stake);
     }
 
-    /**
-     * Update stake - called from rewards contract
-     */
-    function setStake(address _token, address _user, uint _amount, uint256 _stakeId) external {
+    function setStake(
+        address _token,
+        address _user,
+        uint256 _amount,
+        uint256 _stakeId
+    ) external {
         Stake storage stake = Stakes[_user][_token][_stakeId];
 
         stake.amount += _amount; // ? how can I send transfer the amount from the user? what about the previous amount?
     }
 
-    function withdrawToken(address _token, uint256 _amount, address _destination) external onlyOwner {
+    function withdrawToken(
+        address _token,
+        uint256 _amount,
+        address _destination
+    ) external onlyOwner {
         Pool storage pool = Pools[_token];
-        
+
         require(_amount > 0 && pool.totalPooled >= _amount);
-        
+
         // withdraw the token to user wallet
         IERC20(_token).safeTransferFrom(address(this), _destination, _amount);
